+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Put a space between nested unary `-` calls (#170).
 * Allow `-t` flag to preserve compability with old formatter (#166).
 * Support `--machine` flag for machine-readable output (#164).
+* If no paths are provided, read source from stdin (#165).
 
 # 0.1.3